cast distance to obstacle to int

+++ src/ergodic_exploration/collision.cpp
@@ -191,7 +191,7 @@ bool Collision::checkCell(CollisionConfig& cfg, const GridMap& grid, unsigned in
   if (grid.gridBounds(ci, cj) && !(grid.getCell(ci, cj) < occupied_threshold_))
   {
     // squared distance circle center to obstacle
-    const auto sqrd_obs = (cfg.cx - cj) * (cfg.cx - cj) + (cfg.cy - ci) * (cfg.cy - ci);
+    const auto sqrd_obs = static_cast<int>((cfg.cx - cj) * (cfg.cx - cj) + (cfg.cy - ci) * (cfg.cy - ci));
 
     // update smallest squared distance to obstacle
     if (sqrd_obs < cfg.sqrd_obs || cfg.sqrd_obs == -1)
